+++ src/app/metrics/implementations/__init__.py
@@ -1,8 +1,9 @@
 # Import all metric implementations to trigger registration
 from . import license
-from . import documentation  
-from . import maintainer
-from . import performance
+from . import bus_factor  
+from . import ramp_up_time
+from . import performance_claims
 from . import size
-from . import dataset
+from . import dataset_and_code
 from . import code_quality
+from . import dataset_quality
