+++ src/app/metrics/implementations/bus_factor.py
