+++ src/app/metrics/implementations/dataset_and_code.py
@@ -65,65 +65,3 @@ class DatasetAndCodeScoreMetric(BaseMetric):
             f"and {code_status}. Real implementation will evaluate dataset "
             "quality, code completeness, and ecosystem integration."
         )
-
-
-@register("dataset_quality")
-class DatasetQualityMetric(BaseMetric):
-    """
-    Evaluates dataset reliability, popularity, and quality indicators.
-    
-    This metric focuses specifically on the datasets used for training and evaluation:
-    - Dataset size, diversity, and representativeness
-    - Data quality indicators (completeness, accuracy, consistency)
-    - Dataset popularity and community adoption
-    - Citation count and academic recognition
-    - Data collection methodology and ethics
-    - Bias analysis and fairness considerations
-    
-    Future Implementation Notes:
-    - Analyze dataset statistics from Hugging Face datasets hub
-    - Cross-reference with academic papers and citation counts
-    - Evaluate dataset documentation and metadata completeness
-    - Check for bias analysis and ethical considerations
-    - Assess dataset maintenance and update frequency
-    """
-    
-    name = "dataset_quality"
-    
-    def _compute_score(self, resource: ResourceBundle) -> float:
-        """
-        Evaluate quality and reliability of associated datasets.
-        
-        TODO: Replace with real implementation that:
-        1. Fetches dataset metadata from Hugging Face datasets API
-        2. Analyzes dataset size, diversity, and completeness metrics
-        3. Evaluates dataset popularity and community adoption
-        4. Checks for bias analysis and ethical documentation
-        5. Cross-references with academic citations and usage
-        
-        Args:
-            resource: Bundle with dataset URLs to analyze
-            
-        Returns:
-            Score from 0-1 where 1 = high-quality, well-documented, widely-used datasets
-        """
-        if not resource.dataset_urls:
-            return 0.2  # Low score for missing datasets
-            
-        # PLACEHOLDER: Real implementation will analyze actual dataset quality metrics
-        base_score = stable_01(resource.model_url, "dataset_quality")
-        
-        # Bonus for having multiple datasets (better evaluation coverage)
-        dataset_diversity_bonus = min(0.2, len(resource.dataset_urls) * 0.1)
-        
-        return min(1.0, base_score + dataset_diversity_bonus)
-    
-    def _get_computation_notes(self, resource: ResourceBundle) -> str:
-        if not resource.dataset_urls:
-            return f"No datasets found for {resource.model_url}. Quality assessment not possible."
-            
-        return (
-            f"Placeholder dataset quality analysis for {resource.model_url} "
-            f"with {len(resource.dataset_urls)} associated datasets. "
-            "Real implementation will evaluate dataset metrics, popularity, and documentation."
-        )
